+++ core/src/main/java/com/facebook/ktfmt/format/RedundantElementManager.kt
@@ -20,6 +20,7 @@ import org.jetbrains.kotlin.com.intellij.psi.PsiElement
 import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.kdoc.psi.impl.KDocImpl
 import org.jetbrains.kotlin.psi.KtElement
+import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.psi.KtImportList
 import org.jetbrains.kotlin.psi.KtPackageDirective
 import org.jetbrains.kotlin.psi.KtReferenceExpression
@@ -33,8 +34,8 @@ import org.jetbrains.kotlin.psi.psiUtil.startOffset
  */
 object RedundantElementManager {
   /** Remove extra semicolons and unused imports, if enabled in the [options] */
-  fun dropRedundantElements(code: String, options: FormattingOptions): String {
-    val file = Parser.parse(code)
+  internal fun dropRedundantElements(file: KtFile, options: FormattingOptions): String {
+    val code = file.text
     val redundantImportDetector = RedundantImportDetector(enabled = options.removeUnusedImports)
     val redundantSemicolonDetector = RedundantSemicolonDetector()
     val trailingCommaDetector = TrailingCommas.Detector()
@@ -92,12 +93,12 @@ object RedundantElementManager {
     return result.toString()
   }
 
-  fun addRedundantElements(code: String, options: FormattingOptions): String {
+  internal fun addRedundantElements(file: KtFile, options: FormattingOptions): String {
     if (!options.manageTrailingCommas) {
-      return code
+      return file.text
     }
 
-    val file = Parser.parse(code)
+    val code = file.text
     val trailingCommaSuggestor = TrailingCommas.Suggestor()
 
     file.accept(